video_summary/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
google-cloud-aiplatform==1.56.0
google-cloud-storage==2.12.0
